requirements.txt
streamlit
openai-whisper
moviepy
pdf2image
torch
pymupdf